+++ src/main/java/frc/team1778/robot/autonomous/AutoPaths.java
@@ -10,5 +10,7 @@ import jaci.pathfinder.Waypoint;
  */
 public class AutoPaths {
   public static Waypoint[] testPath =
-      new Waypoint[] {new Waypoint(0, 0, Pathfinder.d2r(0.0)), new Waypoint(36.0, 36.0, Pathfinder.d2r(90.0))};
+      new Waypoint[] {
+        new Waypoint(0, 0, Pathfinder.d2r(0.0)), new Waypoint(36.0, 36.0, Pathfinder.d2r(90.0))
+      };
 }
